api-service/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
gunicorn==22.0.0
pika==1.3.2
python-consul==1.1.0
celery==5.4.0
requests==2.32.3
PyJWT==2.8.0